exercises 4.1 - 4.22

+++ fullstackopen-part4/blog/controllers/blogs.js
@@ -1,26 +1,11 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-}
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
 
-blogsRouter.post('/', async (request, response) => {
+    const user = request.user
     const body = request.body
-
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
-
-    body.user = user._id
 
     const blog = new Blog({
         title: body.title,
@@ -49,7 +34,16 @@ blogsRouter.get('/:id', async (request, response) => {
     response.json(blog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+
+    const user = request.user
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() !== user._id.toString()) {
+        return response.status(401).json({ error: 'user can only delete their own post' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
